Moved PackageGraph extension to Workspace.
